Provide to pull/push from a mirror registry
The DOCKER_ENCAPS_IMG_MIRROR command allows to provide a (local/faster)
mirror for the DOCKER_ENCAPS_IMG.
It will still try to download from the origin, and consider the mirror
version only as a fallback.

If the version on the mirror doesn't match the source, it is possible
to set DOCKER_ENCAPS_IMG_MIRROR_UPDATE to true as to automatically
update the mirror.
This assumes that the authentification to the mirror has been taken
care of.